core: make validation a simple state machine instead
With a stateful object like before, it was necessary to work
around clobbering from multiple calls to `#validate()`  (f5dff46e6860732ded366b6cff0087909ee9b79a).

On top of that, `#validate()` was unable to return the result
from when its call happened, it'd either have to wait for the
last validator to complete (thus changing the state away from
pending), or to resolve to a pending state.